changed button style and username display to first and last name

+++ public/js/userprofile.js
@@ -14,44 +14,6 @@ formEl.addEventListener("submit", (event) => {
     saveList();
 });
 
-
-// submitTaskButton.addEventListener("click", (event) => {
-
-//     const taskId = 
-// });
-
-
-// const getList = (id) =>
-//     fetch(`/api/lists/users/${id}`, {
-//         method: 'GET',
-//         headers: {
-//             'Content-Type': 'application/json',
-//         },
-//     });
-
-// const saveList = () => {
-//     const inputListNameText = inputListName.value;
-//     console.log("Checking what we will send to express route")
-//     console.log(inputListNameText)
-
-//     fetch('/api/lists', {
-//         method: 'POST',
-//         // headers: {
-//         //     'Content-Type': 'application/json',
-//         // },
-//         body: JSON.stringify({list_name: inputListNameText}),
-//     }).then(response=>response.json())
-//     .then(data=>{
-//         console.log(data);
-//         // W3school
-//         // var liEl = document.createElement("li")
-//         // liEl.classList.add("some-class1")
-//         // liEl.classList.add("some-class2")
-//         // liEl.textContent = inputListName.value
-//         // ulEl.appendChild(liEl)
-//     })
-// }
-
 const getList = (id) =>
     fetch(`/api/lists/users/${id}`, {
         method: 'GET',
@@ -60,7 +22,7 @@ const getList = (id) =>
         },
     }).then(response => response.json())
         .then(data => {
-            console.log('Task form data',data);
+            console.log('Task form data', data);
             console.log('Name of the task', inputListNameText);
 
             const tableBody = document.querySelector('#tbody');
@@ -73,42 +35,36 @@ const getList = (id) =>
             tableBody.appendChild(newRow);
         })
 
-        const getTasks = (id) =>
-        fetch(`/api/lists`, {
-            method: 'GET',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-        }).then(response => response.json())
-            .then(data => {
-                console.log('Tasks data',data);
-                
-    
-                const tableBody = document.querySelector('#task-table tbody');
-                tableBody.innerHTML = "";
-                
-                data.tasks.forEach(task => {
+const displayList = (id) =>
+    fetch(`/api/lists`, {
+        method: 'GET',
+        headers: {
+            'Content-Type': 'application/json',
+        },
+    }).then(response => response.json())
+        .then(data => {
+            console.log('Tasks data', data);
 
 
-                    
-                    
+            const tableBody = document.querySelector('#task-table tbody');
+            tableBody.innerHTML = "";
 
-                    if(task.list_body){
+            data.tasks.forEach(task => {
+                if (task.list_body) {
 
-                        
-                        let checkboxId = `check-box${task.id}`;
+
+                    let checkboxId = `check-box${task.id}`;
                     const row = document.createElement("tr");
                     row.innerHTML = ` <th scope="row"><input class="todo__checkbox" type="checkbox" id="${checkboxId}"></th>
-                    <td>${task.list_body}</td>
+                    <td class="text-break">${task.list_body}</td>
                     `;
 
                     tableBody.appendChild(row);
-                    }
-                })
-                
+                    // deleteList(checkboxId);
+                }
+
             })
-    
-    
+        })
 
 const saveList = () => {
     const inputListNameText = inputListName.value;
@@ -127,80 +83,33 @@ const saveList = () => {
     })
         .then(data => {
 
-            // //const tableBody = document.querySelector('#tbody');
-            // //const newRow = document.createElement('tr');
-            // const listCell = document.querySelector('.task-name');
-
-            // listCell.innerHTML = data.list_name;
-            // //newRow.appendChild(listCell);
-            // //tableBody.appendChild(newRow);
-
-            getTasks();
+            displayList();
         })
+}
+// need to GET and if it's checked then capture ID 
+const checkedItemBtn = document.querySelector('#btn-complete');
+const deleteList = (id) => {
+
+    const checkedItem = document.querySelector('.todo__checkbox');
+
+    console.log("this is checked item:", id);
+
+    if (checkedItem === on) {
+        fetch(`/api/lists/${id}`, {
+            method: 'DELETE',
+            headers: {
+                'Content-Type': 'application/json',
+            },
+        });
+    }
+    console.log("clicked")
 }
 
 
-const deleteList = (id) =>
-    fetch(`/api/lists/${id}`, {
-        method: 'DELETE',
-        // headers: {
-        //     'Content-Type': 'application/json',
-        // },
-    });
 
-const editList = (list) =>
-    fetch(`/api/lists/${id}`, {
-        method: 'PUT',
-        // headers: {
-        //     'Content-Type': 'application/json',
-        // },
-        body: JSON.stringify(list),
-    });
+checkedItemBtn.addEventListener("click", deleteList);
 
-const renderActiveList = () => {
-    if (activeList.id) {
-        listDescription.setAttribute('readonly', true);
-        listDescription.value = activeList.list_body;
-    } else {
-        listDescription.removeAttribute('readonly');
-        listDescription.value = '';
-    }
-};
-
-const handleListSave = () => {
-    const newList = {
-        list_body: listDescription.value,
-    };
-    saveList(newList).then(() => {
-        getAndRenderList();
-        renderActiveList();
-    });
-};
-
-const handleListDelete = (e) => {
-    // Prevents the click listener for the list from being called when the button inside of it is clicked
-    e.stopPropagation();
-
-    const list = e.target;
-    // const listId = JSON.parse(list.parentElement.getAttribute('not sure what goes here')).id;
-
-    if (activeList.id === listId) {
-        activeList = {};
-    }
-
-    deleteList(listId).then(() => {
-        getAndRenderList();
-        renderActiveList();
-    });
-};
-
-const handleListView = (e) => {
-    e.preventDefault();
-    activeNote = JSON.parse(e.target.parentElement.getAttribute('data-note'));
-    renderActiveNote();
-};
-
-getTasks();
+displayList();
 
 
 
@@ -220,67 +129,3 @@ getTasks();
 
 
 
-//create new list --emily created this
-// const newListHandler = async (event) => {
-//     event.preventDefault();
-
-//     const list_name = document.querySelector('#list_name').value.trim();
-//     if (list_name) {
-//         const response = await fetch(`/api/lists`, {
-//             method: 'POST',
-//             body: JSON.stringify({
-//                 list_name,
-//             }),
-//             headers: {
-//                 'Content-Type': 'application/json',
-//             },
-//         });
-
-//         if (response.ok) {
-//             document.location.replace(`/list/${list_name}`);
-//         } else {
-//             alert('Failed to create list');
-//         }
-//     }
-// };
-
-// //delete list
-// const delListHandler = async (e) => {
-//     if (e.target.hasAttribute('tasks')) {
-//         const id = e.target.getAttribute('tasks');
-
-//         const response = await fetch(`/api/tasks/${id}`, {
-//             method: 'DELETE',
-//         });
-
-//         if (response.ok) {
-//             document.location.replace('/profile');
-//         } else {
-//             alert('Failed to delete task');
-//         }
-//     }
-// };
-
-// //add list row to table
-// const addListHandler = async (e) => {
-
-//     document.getElementById('list-btn').addEventListener('click', function () {
-//         const listName = document.querySelector('list-name').value
-//         const listTable = document.querySelector('listTable')
-
-//         for (let i = 1; i <= listName, i++;) {
-//             // const result = listName;
-//             let newListName = listTable.insertRow(-1);
-
-//             newListName.innerHTML = listName;
-//         }
-//     });
-// };
-
-
-// document
-//     .querySelector('')
-//     .addEventListener('submit', addListHandler);
-// document
-//     .querySelector('')
-//     .addEventListener('submit', delListHandler);
